Run ctags on all gems in the bundle
The `gem ctags` command can do this itself under some circumstances, but
I've found that it doesn't do it when I'm in a devcontainer with bundled
gems stored in /bundle/vendor. I have a hunch that this is because
they're not installed in a location with `gem list` can find them.

This function of mine feels like a quick hack/workaround; there's
probably a better way of getting bundled gems' tag files generated when
bundler runs. I'll need to remember to re-run this as I'm modifying the
bundle, and it's not (currently at least) going to be available by
default in my shell.