Persist the value of BUNDLE_PATH for vim-bundler
The vim-bundler plugin sets Vim's tags option to a comma separated list
of all the tags files in the project's Ruby bundle. It can only do that
if it can read the path to the bundle from the config file though; it
doesn't support reading it from the BUNDLE_PATH environment variable.

So this patch takes the environment variable (if it's set) and writes it
to the config file.